Enhance error handling in product and review repositories
- Added checks to ensure the existence of products and reviews before performing operations, throwing NotFoundError when not found.
- Introduced a new method in the ReviewRepository to check for review existence, improving error handling and data integrity.
- These changes enhance the robustness of the product and review management systems by preventing operations on non-existent entities.

+++ backend/src/db/repos/products/product.repo.ts
@@ -18,6 +18,9 @@ export class ProductRepository {
     productId: string,
     userId: string
   ): Promise<boolean> {
+    if (!(await this.checkProductExists(productId)))
+      throw new NotFoundError('Product not found');
+
     const product = await this.collection.findOne(
       { _id: new ObjectId(productId) },
       { projection: { user: 1 } }
@@ -53,6 +56,9 @@ export class ProductRepository {
     productId: string,
     productData: Partial<Omit<Product_DTO, 'id' | 'createdAt' | 'user'>>
   ): Promise<Product_DTO | null> {
+    if (!(await this.checkProductExists(productId)))
+      throw new NotFoundError('Product not found');
+
     const updatedProduct = await this.collection.findOneAndUpdate(
       { _id: new ObjectId(productId) },
       { $set: { ...productData, updatedAt: new Date() } },
@@ -86,6 +92,9 @@ export class ProductRepository {
     productId: string,
     session?: ClientSession
   ): Promise<boolean> {
+    if (!(await this.checkProductExists(productId)))
+      throw new NotFoundError('Product not found');
+
     const result = await this.collection.deleteOne(
       {
         _id: new ObjectId(productId),
@@ -95,15 +104,13 @@ export class ProductRepository {
     return result.deletedCount === 1;
   }
 
-  /**
-   * Updates the average rating of a product based on its reviews.
-   * @param productId - The ID of the product.
-   * @param session - (Optional) The MongoDB session for transactions.
-   */
   public async updateAverageRating(
     productId: string,
     session?: ClientSession
   ): Promise<void> {
+    if (!(await this.checkProductExists(productId)))
+      throw new NotFoundError('Product not found');
+
     const pipeline = [
       {
         $match: { product: new ObjectId(productId) },
